// client/src/app/Models/usermaintenance.ts
export interface IUsermaintenance {
    UserId: number;
    UserCode: number;

    UserName: string;

    Email: string;
    Status:string;
    StatusID: number;
    Description: string;
    StausType:string;
    CreatedBy: number;
    CreatedDate: any;
    ModifiedBy: number;
    ModifiedDate: any;

    CompanyID:number;
    Password: string;
    Active: number;

    InActive:number;
    Company:string;
    
}